created Hooks to dynamically check for month/year for the statistics Labels
