+++ src/Hooks/DateUtils.tsx
@@ -1,0 +1,27 @@
+export function getCurrentMonthName() {
+  const months = [
+    'January',
+    'February',
+    'March',
+    'April',
+    'May',
+    'June',
+    'July',
+    'August',
+    'September',
+    'October',
+    'November',
+    'December',
+  ];
+
+  const currentDate = new Date();
+  const currentMonth = months[currentDate.getMonth()];
+
+  return currentMonth;
+}
+
+export function getCurrentYear() {
+  const currentDate = new Date();
+  const currentYear = currentDate.getFullYear();
+  return currentYear.toString();
+}
